[SPARK-31176][SQL] Remove support for 'e'/'c' as datetime pattern charactar
