+++ sql/core/src/test/resources/sql-tests/results/datetime.sql.out
@@ -1,5 +1,5 @@
 -- Automatically generated by SQLQueryTestSuite
--- Number of queries: 68
+-- Number of queries: 73
 
 
 -- !query
@@ -559,3 +559,43 @@ select to_timestamp("12.1234019-10-06S10:11", "ss.SSSSy-MM-dd'S'HH:mm")
 struct<to_timestamp('12.1234019-10-06S10:11', 'ss.SSSSy-MM-dd\'S\'HH:mm'):timestamp>
 -- !query output
 0019-10-06 10:11:12.1234
+
+
+-- !query
+select to_timestamp("2019-10-06S", "yyyy-MM-dd'S'")
+-- !query schema
+struct<to_timestamp('2019-10-06S', 'yyyy-MM-dd\'S\''):timestamp>
+-- !query output
+2019-10-06 00:00:00
+
+
+-- !query
+select to_timestamp("S2019-10-06", "'S'yyyy-MM-dd")
+-- !query schema
+struct<to_timestamp('S2019-10-06', '\'S\'yyyy-MM-dd'):timestamp>
+-- !query output
+2019-10-06 00:00:00
+
+
+-- !query
+select to_timestamp("2019-10-06 Sunday", "yyyy-MM-dd eeee")
+-- !query schema
+struct<to_timestamp('2019-10-06 Sunday', 'yyyy-MM-dd eeee'):timestamp>
+-- !query output
+NULL
+
+
+-- !query
+select to_timestamp("2019-10-06 Sunday", "yyyy-MM-dd cccc")
+-- !query schema
+struct<to_timestamp('2019-10-06 Sunday', 'yyyy-MM-dd cccc'):timestamp>
+-- !query output
+NULL
+
+
+-- !query
+select to_timestamp("2019-10-06 Sunday", "yyyy-MM-dd uuuu")
+-- !query schema
+struct<to_timestamp('2019-10-06 Sunday', 'yyyy-MM-dd uuuu'):timestamp>
+-- !query output
+2019-10-06 00:00:00
